new my_put_nbr

+++ include/my.h
@@ -10,7 +10,7 @@
 int my_put_nbr(int nb);
 void my_putchar(char c);
 int my_putstr(char const *str);
-int flag_f(float nb);
+int flag_f(double nb);
 int flag_s(char *str);
 int flag_e(float nb);
 #endif /* MY_H */
